# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovaRoute API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/student/profile": {
      "get": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Profile",
        "operationId": "profile_api_student_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Edit Profile",
        "operationId": "edit_profile_api_student_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/profile/clear-resume": {
      "post": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Clear Derived",
        "operationId": "clear_derived_api_student_profile_clear_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Version"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/employer/profile": {
      "get": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Company",
        "operationId": "company_api_employer_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Edit Company",
        "operationId": "edit_company_api_employer_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload-resume": {
      "post": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Upload",
        "operationId": "upload_api_upload_resume_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_upload_resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/student/resumes/{identifier}": {
      "get": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Resume Detail",
        "operationId": "resume_detail_api_student_resumes__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Delete Resume",
        "operationId": "delete_resume_api_student_resumes__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/resumes/{identifier}/download": {
      "get": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Download",
        "operationId": "download_api_student_resumes__identifier__download_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/resumes/{identifier}/confirm": {
      "post": {
        "tags": [
          "Profiles and resumes"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_student_resumes__identifier__confirm_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_upload_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_upload_resume_post"
      },
...
      "CompanyOut": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "website": {
            "type": "string",
            "title": "Website"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status",
            "default": "pending"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "demo_company": {
            "type": "boolean",
            "title": "Demo Company",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "website",
          "industry",
          "location",
          "description"
        ],
        "title": "CompanyOut"
      },
      "CompanyPatch": {
        "properties": {
          "company_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Company Name"
          },
          "website": {
            "type": "string",
            "maxLength": 500,
            "title": "Website",
            "default": ""
          },
          "industry": {
            "type": "string",
            "maxLength": 150,
            "title": "Industry",
            "default": ""
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "title": "Location",
            "default": ""
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "company_name"
        ],
        "title": "CompanyPatch"
      },
...
      "Extraction": {
        "properties": {
          "skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Skills"
          },
          "coursework": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Coursework"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Projects"
          },
          "education": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 10,
            "title": "Education"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "skills",
          "coursework",
          "projects",
          "education"
        ],
        "title": "Extraction"
      },
      "ExtractionConfirm": {
        "properties": {
          "skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Skills"
          },
          "coursework": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Coursework"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Projects"
          },
          "education": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 10,
            "title": "Education"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "skills",
          "coursework",
          "projects",
          "education",
          "version"
        ],
        "title": "ExtractionConfirm"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "OK": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OK"
      },
...
      "Preferences": {
        "properties": {
          "work_mode": {
            "type": "string",
            "enum": [
              "",
              "remote",
              "hybrid",
              "onsite"
            ],
            "title": "Work Mode",
            "default": ""
          },
          "interests": {
            "type": "string",
            "maxLength": 500,
            "title": "Interests",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Preferences"
      },
      "ProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "extracted_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Skills"
          },
          "coursework": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Coursework"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "title": "Projects"
          },
          "degree": {
            "type": "string",
            "title": "Degree"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "preferences": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Preferences"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "extracted_skills",
          "coursework",
          "projects",
          "degree",
          "location",
          "preferences",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileOut"
      },
      "ProfilePatch": {
        "properties": {
          "extracted_skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Extracted Skills"
          },
          "coursework": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Coursework"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Projects"
          },
          "degree": {
            "type": "string",
            "maxLength": 200,
            "title": "Degree"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "title": "Location"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "extracted_skills",
          "coursework",
          "projects",
          "degree",
          "location",
          "version"
        ],
        "title": "ProfilePatch"
      },
      "Project": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 1500,
            "title": "Description"
          },
          "technologies": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Technologies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description",
          "technologies"
        ],
        "title": "Project"
      },
...
      "ResumeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "parse_state": {
            "type": "string",
            "title": "Parse State"
          },
          "extracted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Extraction"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              }
            ],
            "title": "Extracted"
          },
          "failure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure"
          },
          "profile_version": {
            "type": "integer",
            "title": "Profile Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "confirmation_needed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Needed"
          },
          "ai_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Mode"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "filename",
          "content_hash",
          "parse_state",
          "extracted",
          "failure",
          "profile_version",
          "created_at",
          "updated_at"
        ],
        "title": "ResumeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Version": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "Version"
      }
    }
  }
}